requirements.txt
aiofiles==23.2.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
msoffcrypto-tool==5.1.1
numpy==1.26.0
olefile==0.46
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pefile==2023.2.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyotp==2.9.0
pypi-sqlite-cipher==1.0.1
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2021.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uvicorn==0.23.2